Miriup, fix example warnings, move most routes from main.rs to render/mod.rs
