+++ Cargo.toml
@@ -18,7 +18,7 @@ regex = "0.2.5"
 syntect = "2.0.0"
 lazy_static = "1.0.0"
 cgraph = { git = "https://github.com/oli-obk/cgraph.git", rev = "b65e460ee323b31dca55b5541141a6b73272e72a" }
-miri = { git = "https://github.com/solson/miri.git", branch = "rustup" }
+miri = { git = "https://github.com/solson/miri.git" }
 rocket = "0.3.8"
 rocket_codegen = "0.3.8"
